pyproject.toml
[project]
name = "railway-station-api"
version = "1.0.0"
description = "A complex train management system for booking trips, managing crew, and stations efficiently"
authors = [
    {name = "Artem",email = "stilon670@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.1.7,<6.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "markdown (>=3.7,<4.0)",
    "django-debug-toolbar (>=5.0.1,<6.0.0)",
    "black[d] (>=25.1.0,<26.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "psycopg-binary (>=3.2.6,<4.0.0)",
    "psycopg (>=3.2.6,<4.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "pytz (>=2025.1,<2026.0)",
    "flake8 (>=7.1.2,<8.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
